paper_trading/requirements.txt
# Core Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0
redis>=5.0.0
pyyaml>=6.0.0

# Web Framework
flask>=2.3.0
flask-socketio>=5.3.0

# Data Processing
python-binance>=1.0.19
ccxt>=4.0.0
ta>=0.10.0
ta-lib>=0.4.0
yfinance>=0.2.0
websockets>=11.0.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
optuna>=3.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Visualization
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0 